+++ NEWS.md
@@ -8,6 +8,7 @@
 * Replace the `iris` dataset with the `airquality` dataset in all documentation, examples, and tests (#1271).
 * Assign functions created with `code_to_function()` to the proper environment (#1275, @robitalec).
 * Store tracebacks as character vectors and restrict the contents of error objects to try to prevent accidental storage of large data from the environment (#1276, @billdenney).
+* Strongly depend on `tidyselect` (#1274, @dernst).
 
 # Version 7.12.2