+++ NEWS.md
@@ -1,4 +1,4 @@
-# Version 7.12.2.9000
+# Version 7.12.3
 
 ## Bug fixes